design: 마이페이지 관심 목록 리스트 생성
