+++ abandonedpets/src/pages/Mypage.tsx
@@ -1,0 +1,13 @@
+import NavBar from "../components/NavBar";
+import FavoriteList from "../components/mypage/FavoriteList";
+
+function Mypage() {
+    return (
+        <>
+            <NavBar />
+            <FavoriteList />
+        </>
+    )
+}
+
+export default Mypage;
